feat : Image Picker 구현 및 코드 정리
