+++ aljo/Projects/UserInterface/ASAPKit/Sources/Controllers/ASImagePickerViewController/ASImagePickerViewController.swift
@@ -10,95 +10,6 @@ import Photos
 
 import SnapKit
 
-final class AssetImageGridCell: UICollectionViewCell {
-  static let identifier = "AssetImageGridCell"
-  // View Properties
-  private let imageView: UIImageView = {
-    let imageView = UIImageView()
-    imageView.clipsToBounds = true
-    imageView.contentMode = .scaleAspectFill
-    return imageView
-  }()
-  
-  private let selectedBlurView: UIView = {
-    let view = UIView()
-    view.backgroundColor = .black01
-    view.alpha = .zero
-    return view
-  }()
-  
-  private let checkButton: UIButton = {
-    var configuration = UIButton.Configuration.filled()
-    configuration.baseBackgroundColor = .white
-    configuration.image = .Icon.check_small.withTintColor(.white)
-    configuration.cornerStyle = .capsule
-    
-    let button = UIButton(configuration: configuration)
-    return button
-  }()
-  
-  // Life Cycle
-  override func layoutSubviews() {
-    super.layoutSubviews()
-    
-    configureUI()
-  }
-  
-  override func prepareForReuse() {
-    super.prepareForReuse()
-    
-    self.imageView.image = nil
-    setSelected(to: false)
-  }
-  
-  // Configure Method
-  func setImage(to image: UIImage?) {
-    self.imageView.image = image
-  }
-  
-  func setSelected(to isSelected: Bool, with color: UIColor = .red01) {
-    self.layer.borderColor = isSelected ? color.cgColor : UIColor.clear.cgColor
-    self.layer.borderWidth = 2
-    self.checkButton.configuration?.baseBackgroundColor = isSelected ? color : .white
-//    self.selectedBlurView.alpha = isSelected ? 0.5 : 0
-    self.checkButton.isSelected = isSelected
-  }
-  
-  func getImage() -> UIImage? {
-    return imageView.image
-  }
-}
-
-// Configure UI Method
-private extension AssetImageGridCell {
-  func configureUI() {
-    configureHierarchy()
-    makeConstraints()
-  }
-  
-  func configureHierarchy() {
-    contentView.addSubview(imageView)
-    contentView.addSubview(selectedBlurView)
-    contentView.addSubview(checkButton)
-  }
-  
-  func makeConstraints() {
-    imageView.snp.makeConstraints {
-      $0.edges.equalToSuperview()
-    }
-    
-    selectedBlurView.snp.makeConstraints {
-      $0.edges.equalTo(imageView)
-    }
-    
-    checkButton.snp.makeConstraints {
-      $0.top.equalToSuperview().offset(10)
-      $0.trailing.equalToSuperview().offset(-10)
-      $0.width.height.equalTo(20)
-    }
-  }
-}
-
 public protocol ASImagePickerDelegate: AnyObject {
   func imagePicker(_ picker: ASImagePickerViewController, didComplete images: [UIImage?])
 }
@@ -162,21 +73,16 @@ public class ASImagePickerViewController: UIViewController {
   }
   
   private func fetchSelectedImages(completionHandler: @escaping ([UIImage?]) -> Void) {
-    let manager = PHImageManager.default()
-    let option = PHImageRequestOptions()
-    option.deliveryMode = .highQualityFormat
-    option.resizeMode = .fast
+
     var images: [UIImage?] = []
-    print(UIScreen.main.scale)
+    let frame = view.frame
+    let size = CGSize(width: frame.width, height: frame.height)
     let group = DispatchGroup()
     
     for asset in selectedItems {
-      let size = CGSize(width: asset.pixelWidth, height: asset.pixelHeight)
-      DispatchQueue.global().async(group: group) {
+      DispatchQueue.global().async(group: group) { [weak self] in
         group.enter()
-        manager.requestImage(
-          for: asset, targetSize: size, contentMode: .aspectFit, options: option
-        ) { image, _ in
+        self?.fetchItem(to: asset, with: size) { image in
           images.append(image)
           group.leave()
         }
@@ -186,6 +92,21 @@ public class ASImagePickerViewController: UIViewController {
     group.notify(queue: .main) {
       completionHandler(images)
     }
+  }
+  
+  private func fetchItem(
+    to asset: PHAsset,
+    with size: CGSize,
+    completion: @escaping (UIImage?) -> Void
+  ) {
+    let manager = PHImageManager.default()
+    let option = PHImageRequestOptions()
+    option.deliveryMode = .highQualityFormat
+    option.resizeMode = .fast
+    
+    manager.requestImage(
+      for: asset, targetSize: size, contentMode: .aspectFit, options: option
+    ) { image, _ in completion(image) }
   }
 }
 
@@ -205,19 +126,10 @@ extension ASImagePickerViewController: UICollectionViewDataSource {
     guard let cell = collectionView.dequeueReusableCell(
       withReuseIdentifier: AssetImageGridCell.identifier,
       for: indexPath
-    ) as? AssetImageGridCell else {
-      return AssetImageGridCell()
-    }
-    let asset = photos?.object(at: indexPath.item) ?? PHAsset()
+    ) as? AssetImageGridCell else { return AssetImageGridCell() }
     
-    PHImageManager.default().requestImage(
-      for: asset,
-      targetSize: thumbnailSize,
-      contentMode: .aspectFill,
-      options: nil
-    ) { image, _ in
-      cell.setImage(to: image)
-    }
+    let asset = photos?.object(at: indexPath.item) ?? PHAsset()
+    fetchItem(to: asset, with: thumbnailSize) { cell.setImage(to: $0) }
     
     return cell
   }
@@ -258,17 +170,13 @@ extension ASImagePickerViewController: UICollectionViewDelegateFlowLayout {
     _ collectionView: UICollectionView,
     layout collectionViewLayout: UICollectionViewLayout,
     minimumLineSpacingForSectionAt section: Int
-  ) -> CGFloat {
-    return 1.5
-  }
+  ) -> CGFloat { return 1.5 }
   
   public func collectionView(
     _ collectionView: UICollectionView,
     layout collectionViewLayout: UICollectionViewLayout,
     minimumInteritemSpacingForSectionAt section: Int
-  ) -> CGFloat {
-    return 1.5
-  }
+  ) -> CGFloat { return 1.5 }
 }
 
 private extension ASImagePickerViewController {
@@ -291,7 +199,7 @@ private extension ASImagePickerViewController {
       DispatchQueue.main.async {
         switch status {
           case .denied, .restricted:
-            print(status)
+            debugPrint("[에러] 사용자가 개인정보 처리 방침을 거부했습니다.")
           default:
             return
         }
@@ -309,36 +217,32 @@ extension ASImagePickerViewController: PHPhotoLibraryChangeObserver {
     DispatchQueue.main.sync {
       self.photos = changes.fetchResultAfterChanges
       
-      if changes.hasIncrementalChanges == true {
-        collectionView.performBatchUpdates {
-          if let removed = changes.removedIndexes,
-             removed.isEmpty == false {
-            let indexPaths = removed.map { IndexPath(item: $0, section: .zero) }
-            collectionView.deleteItems(at: indexPaths)
-          }
-          
-          if let inserted = changes.insertedIndexes,
-             inserted.isEmpty == false {
-            let indexPaths = inserted.map { IndexPath(item: $0, section: .zero)}
-            collectionView.insertItems(at: indexPaths)
-          }
-          
-          changes.enumerateMoves {
-            let fromIndexPath = IndexPath(item: $0, section: .zero)
-            let toIndexPath = IndexPath(item: $1, section: .zero)
-            self.collectionView.moveItem(at: fromIndexPath, to: toIndexPath)
-          }
-          
-          if let changed = changes.changedIndexes,
-             changed.isEmpty == false {
-            let indexPaths = changed.map { IndexPath(item: $0, section: .zero) }
-            collectionView.reloadItems(at: indexPaths)
-          }
-        }
-      }
-      
       if changes.hasIncrementalChanges == false {
         collectionView.reloadData()
+        return
+      }
+      
+      collectionView.performBatchUpdates {
+        if let removed = changes.removedIndexes, removed.isEmpty == false {
+          let indexPaths = removed.map { IndexPath(item: $0, section: .zero) }
+          collectionView.deleteItems(at: indexPaths)
+        }
+        
+        if let inserted = changes.insertedIndexes, inserted.isEmpty == false {
+          let indexPaths = inserted.map { IndexPath(item: $0, section: .zero) }
+          collectionView.insertItems(at: indexPaths)
+        }
+        
+        changes.enumerateMoves { [weak self] in
+          let fromIndexPath = IndexPath(item: $0, section: .zero)
+          let toIndexPath = IndexPath(item: $1, section: .zero)
+          self?.collectionView.moveItem(at: fromIndexPath, to: toIndexPath)
+        }
+        
+        if let changed = changes.changedIndexes, changed.isEmpty == false {
+          let indexPaths = changed.map { IndexPath(item: $0, section: .zero) }
+          collectionView.reloadItems(at: indexPaths)
+        }
       }
     }
   }
@@ -360,7 +264,6 @@ private extension ASImagePickerViewController {
     }
     
     let confirmAction = UIAction { [weak self] _ in
-      // Delegate Method
       guard let self = self else { return }
       fetchSelectedImages { images in
         self.delegate?.imagePicker(self, didComplete: images)
@@ -389,6 +292,7 @@ private extension ASImagePickerViewController {
   }
 }
 
+// TODO: - Extension으로 빼기
 extension UINavigationController {
   func setUpAppearance() {
     let appearance = UINavigationBarAppearance()
